+++ packages/unified-latex-to-pretext/tests/break-on-boundaries.test.ts
@@ -2,7 +2,7 @@ import { describe, it, expect } from "vitest";
 import util from "util";
 import { getParser } from "@unified-latex/unified-latex-util-parse";
 import { printRaw } from "@unified-latex/unified-latex-util-print-raw";
-import { breakOnBoundaries } from "../libs/break-on-boundaries";
+import { breakOnBoundaries } from "../libs/pre-conversion-subs/break-on-boundaries";
 
 // Make console.log pretty-print by default
 const origLog = console.log;
